refactor: extract k8s subnet discovery logic into a package
