+++ apiserver/restrict_caasmodel_test.go
@@ -40,6 +40,25 @@ func (s *RestrictCAASModelSuite) TestSpacesReloadAllowed(c *tc.C) {
 	s.assertMethod(c, "Spaces", 6, "ReloadSpaces")
 }
 
+func (s *RestrictCAASModelSuite) TestSpacesReadMethodsAllowed(c *tc.C) {
+	s.assertMethod(c, "Spaces", 6, "ListSpaces")
+	s.assertMethod(c, "Spaces", 6, "ShowSpace")
+}
+
+func (s *RestrictCAASModelSuite) TestSpacesMutationMethodsNotAllowed(c *tc.C) {
+	for _, method := range []string{
+		"CreateSpaces",
+		"MoveSubnets",
+		"RemoveSpace",
+		"RenameSpace",
+	} {
+		caller, err := s.root.FindMethod("Spaces", 6, method)
+		c.Check(err, tc.ErrorMatches, `facade method "Spaces\.`+method+`" not supported on container models`)
+		c.Check(err, tc.ErrorIs, errors.NotSupported)
+		c.Check(caller, tc.IsNil)
+	}
+}
+
 func (s *RestrictCAASModelSuite) TestNotAllowed(c *tc.C) {
 	caller, err := s.root.FindMethod("Firewaller", 1, "WatchOpenedPorts")
 	c.Assert(err, tc.ErrorMatches, `facade "Firewaller" not supported on container models`)
